ML/requirements.txt
# ML Pattern Trading System Dependencies

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Technical analysis
TA-Lib>=0.4.28
pandas-ta>=0.3.14b

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0
scipy>=1.11.0

# Data fetching
yfinance>=0.2.28
nsepy>=0.8
requests>=2.31.0

# Database
sqlalchemy>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Progress tracking
tqdm>=4.65.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Logging
colorlog>=6.7.0

# Web framework (for review interface)
flask>=3.0.0
flask-cors>=4.0.0

# Configuration
python-dotenv>=1.0.0

# Multi-Modal System (NEW!)
# Sentiment Analysis
google-generativeai>=0.3.0

# Price Prediction (FREE alternative to TimeGPT!)
statsforecast>=1.6.0
fugue>=0.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

